Rework Realm Task base model

+++ Source/NexusForever.Database.World/Model/RealmTaskModel.cs
@@ -9,14 +9,14 @@ namespace NexusForever.Database.World.Model
     public class RealmTaskModel
     {
         public uint Id { get; set; }
-        public uint Type { get; set; }
+        public byte Type { get; set; }
         public string Value { get; set; }
         public uint CharacterId { get; set; }
         public uint AccountId { get; set; }
         public uint GuildId { get; set; }
         public uint ReferenceId { get; set; }
         public string ReferenceValue { get; set; }
-        public uint Status { get; set; }
+        public byte Status { get; set; }
         public string StatusDescription { get; set; }
         public DateTime CreateTime { get; set; }
         public string CreatedBy { get; set; }
